Change CI to CircleCI
A recent commit pinned emsdk version to 1.37.28,
which was meant to be built with tools not available
under ubuntu:14.04 (which is the most recent build
box available in travis at this moment). The only
solution to this problem, which happens to also be
the more future proof one, is to build emsdk in the
CI environment instead of relying on prebuilt binaries.
Building emsdk inside travis without a paid plan
is not possible though, because clang takes a long
time to build and travis will abort jobs that take
too long to finish. CircleCI on the other hand,
allows for much longer built times, aside from
offering the ability to create your own base build
container, where emsdk is already prebuilt alongside
all (or almost all) of its dependencies.